tests/wpt/web-platform-tests/credential-management/fedcm-multi-idp/multiple-gets-after-onload.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP multiple gets after onload test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>

<script type="module">
import {set_fedcm_cookie, set_alt_fedcm_cookie, default_request_options, default_alt_request_options} from '../support/fedcm-helper.sub.js';

let cookies_promise = Promise.all([set_fedcm_cookie(), set_alt_fedcm_cookie()]);
const window_loaded = new Promise(resolve => {
  window.addEventListener('load', () => {
    resolve();
  });
});

promise_test(async t => {
  await Promise.all([cookies_promise, window_loaded]);
  const first_cred = navigator.credentials.get(default_request_options());
  const second_cred = navigator.credentials.get(default_alt_request_options());
  // NetworkError is returned when another IDP is selected.
  await promise_rejects_dom(t, 'NetworkError', second_cred);
  const first = await first_cred;
  assert_equals(first.token, "token");
}, "No `get` calls before or during onload, multiple `get` calls after onload in the same task are allowed.");

</script>
